Fix a large amount of PHPCS exceptions
This is mostly about documentation, but not exclusively. In a few
places I found it easier to fix the code instead of working around
an issue.

I'm intentionally going with rather vague "mixed" types in a lot of
places where it's not 100% clear how the code is supposed to behave.

Depends-On: I75178b9a67007a0ac0f9c82ecb2a1d7251086209
Change-Id: I26cc5a978e887c7a4d58ebb1064935abb081f4a7